requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

pymupdf>=1.23.0
faster-whisper>=1.0.0

chromadb>=0.4.0
sentence-transformers>=2.2.0

langchain-text-splitters>=0.2.0

google-generativeai>=0.8.0
